[gradle] Rename app Id with yichunyen.swapi.demo
