+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "yichunyen.ithome2020.challenge"
+        applicationId "yichunyen.swapi.demo"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 10
