+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="yichunyen.ithome2020.challenge">
+    package="yichunyen.swapi.demo">
 
     <uses-permission android:name="android.permission.INTERNET" />
