--- app/src/main/java/yichunyen/ithome2020/challenge/data/FilmResponse.kt
@@ -1,5 +1,0 @@
-package yichunyen.ithome2020.challenge.data
-
-data class FilmResponse(
-    val results: List<Film>
-)
