+++ app/src/main/java/yichunyen/swapi/demo/data/FilmResponse.kt
@@ -1,0 +1,7 @@
+package yichunyen.swapi.demo.data
+
+import yichunyen.swapi.demo.data.Film
+
+data class FilmResponse(
+    val results: List<Film>
+)
