+++ app/src/main/java/yichunyen/swapi/demo/data/Gender.kt
@@ -1,4 +1,4 @@
-package yichunyen.ithome2020.challenge.data
+package yichunyen.swapi.demo.data
 
 /**
  * the gender enum for profile
