+++ app/src/main/java/yichunyen/swapi/demo/network/NetworkManager.kt
@@ -1,4 +1,4 @@
-package yichunyen.ithome2020.challenge.network
+package yichunyen.swapi.demo.network
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -11,14 +11,16 @@ class NetworkManager private constructor() {
         val client: NetworkService
             get() {
                 if (instance == null) {
-                    instance = NetworkManager()
+                    instance =
+                        NetworkManager()
                     retrofit = Retrofit.Builder()
                         .addConverterFactory(GsonConverterFactory.create())
                         .baseUrl(BASE_URL)
                         .build()
                 }
 
-                return retrofit!!.create(NetworkService::class.java)
+                return retrofit!!.create(
+                    NetworkService::class.java)
             }
 
         fun clear() {
